# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trasteros API",
        "description": "API para la gesti\u00f3n de trasteros, contratos, clientes y finanzas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/contratos": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Listar contratos",
                "description": "Obtiene la lista de todos los contratos",
                "operationId": "get_contratos_list",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activo",
                                "pendiente",
                                "finalizado",
                                "cancelado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de contratos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contrato"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Crear contrato",
                "description": "Crea un nuevo contrato",
                "operationId": "post_contratos_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trasteroId",
                                    "clienteId",
                                    "fechaInicio",
                                    "precioMensual"
                                ],
                                "properties": {
                                    "trasteroId": {
                                        "type": "integer"
                                    },
                                    "clienteId": {
                                        "type": "integer"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "fechaFin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31",
                                        "nullable": true
                                    },
                                    "precioMensual": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "fianza": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100,
                                        "nullable": true
                                    },
                                    "fianzaPagada": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contrato creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contrato"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validacion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Trastero ya alquilado o solapamiento de fechas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/contratos/trastero/{trasteroId}": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Contratos por trastero",
                "description": "Obtiene todos los contratos de un trastero especifico",
                "operationId": "get_contratos_by_trastero",
                "parameters": [
                    {
                        "name": "trasteroId",
                        "in": "path",
                        "description": "ID del trastero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "onlyActivos",
                        "in": "query",
                        "description": "Solo contratos activos",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de contratos del trastero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contrato"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contratos/cliente/{clienteId}": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Contratos por cliente",
                "description": "Obtiene todos los contratos de un cliente especifico",
                "operationId": "get_contratos_by_cliente",
                "parameters": [
                    {
                        "name": "clienteId",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    },
                    {
                        "name": "onlyActivos",
                        "in": "query",
                        "description": "Solo contratos activos",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de contratos del cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contrato"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginatedMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contratos/{id}/finalizar": {
            "patch": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Finalizar contrato anticipadamente",
                "description": "Finaliza un contrato anticipadamente estableciendo la fecha de fin a hoy",
                "operationId": "patch_contratos_finalizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del contrato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato finalizado anticipadamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contrato"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contrato no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contratos/{id}/cancelar": {
            "patch": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Cancelar contrato",
                "description": "Marca un contrato como cancelado",
                "operationId": "patch_contratos_cancelar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del contrato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "pattern": "\\d+"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato cancelado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contrato"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contrato no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contrato": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "trasteroId": {
                        "type": "integer"
                    },
                    "clienteId": {
                        "type": "integer"
                    },
                    "fechaInicio": {
                        "type": "string",
                        "format": "date"
                    },
                    "fechaFin": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "precioMensual": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "fianza": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "fianzaPagada": {
                        "type": "boolean"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "activo",
                            "pendiente",
                            "finalizado",
                            "cancelado"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "JWT Authorization header usando el esquema Bearer. Ejemplo: \"Authorization: Bearer {token}\"",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}